PlusBinance/requirements.txt
KekikTaban
KekikSpatula
user-agents
Flask
Flask-Sitemap
waitress
Flask-RESTful
Flask-COMBO-JSONAPI
Flask-Bootstrap
WTForms
Flask_WTF
Flask_Login
pymongo
dnspython
pymongo[srv]
PyJWT
pyTelegramBotAPI
python-binance==0.7.11
deep_translator
typing